Add analytics module, AWS services & IDEA updates
Introduce analytics-service (Kafka Streams topologies, DTOs, JsonSerde, application.yml) and scaffold multiple new services (aws-dynamoDB-service, aws-lambda-service, aws-s3-service, aws-ses-service, mongoDB-service) including Dockerfiles, poms and basic controllers/services. Update IDE project files (.idea/compiler.xml, .idea/misc.xml) to register new modules, annotation processor entries and javac -parameters settings. Also add Helm charts, gateway k8s manifests, prometheus/logstash assets, a local DynamoDB DB file, and miscellaneous updates to common and kafka-common modules.

+++ auth-service/src/main/java/com/cinema/auth/service/AuthService.java
@@ -31,7 +31,7 @@ public class AuthService {
     private final RedisService redisService;
 
     @Value("${jwt.access-token-expire}")
-    private Integer accessTokenExpire;
+    private Long accessTokenExpire;
 
     public JwtResponse login(AuthRequest request) {
         String username = request.username();
